requirements.txt
# LangGraph Agent Testing Demo Dependencies

# Core LangChain ecosystem
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.0.10
langgraph>=0.2.0

# LLM and AI services (for when not using MockLLM)
openai>=1.0.0

# Evaluation and testing
deepeval>=0.3.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-html>=3.0.0

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.0.0

# Development tools (optional)
pytest-xdist>=3.0.0
pytest-rerunfailures>=10.0.0